+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderModalContainer/modals/RunFailedModal.tsx
@@ -1,34 +1,28 @@
 import { useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import { css } from 'styled-components'
 
 import { RUN_STATUS_SUCCEEDED } from '@opentrons/api-client'
 import {
   ALIGN_CENTER,
-  BORDERS,
-  COLORS,
   DIRECTION_COLUMN,
   DIRECTION_ROW,
-  DISPLAY_FLEX,
   Flex,
   Icon,
   JUSTIFY_SPACE_BETWEEN,
   LegacyStyledText,
   Link,
   Modal,
-  OVERFLOW_AUTO,
-  OVERFLOW_WRAP_ANYWHERE,
   PrimaryButton,
   SPACING,
   TYPOGRAPHY,
 } from '@opentrons/components'
 
 import { useDownloadRunLog } from '../../../../hooks'
+import { ErrorContent } from './ErrorContent'
 
 import type { MouseEventHandler } from 'react'
 import type { RunStatus } from '@opentrons/api-client'
 import type { ModalProps } from '@opentrons/components'
-import type { RunCommandError } from '@opentrons/shared-data'
 import type { UseRunErrorsResult } from '../../hooks'
 
 // Note(kk:08/07/2023)
@@ -102,49 +96,6 @@ export function RunFailedModal({
     downloadRunLog()
   }
 
-  interface ErrorContentProps {
-    errors: RunCommandError[]
-    isSingleError: boolean
-  }
-  const ErrorContent = ({
-    errors,
-    isSingleError,
-  }: ErrorContentProps): JSX.Element => {
-    return (
-      <Flex flexDirection={DIRECTION_COLUMN}>
-        <LegacyStyledText as="p" fontWeight={TYPOGRAPHY.fontWeightSemiBold}>
-          {isSingleError
-            ? t('error_info', {
-                errorType: errors[0].errorType,
-                errorCode: errors[0].errorCode,
-              })
-            : runStatus === RUN_STATUS_SUCCEEDED
-            ? t(errors.length > 1 ? 'no_of_warnings' : 'no_of_warning', {
-                count: errors.length,
-              })
-            : t(errors.length > 1 ? 'no_of_errors' : 'no_of_error', {
-                count: errors.length,
-              })}
-        </LegacyStyledText>
-        <Flex css={ERROR_MESSAGE_STYLE}>
-          {' '}
-          {errors.map((error, index) => (
-            <LegacyStyledText
-              as="p"
-              textAlign={TYPOGRAPHY.textAlignLeft}
-              key={index}
-            >
-              {' '}
-              {isSingleError
-                ? error.detail
-                : `${error.errorCode}: ${error.detail}`}
-            </LegacyStyledText>
-          ))}
-        </Flex>
-      </Flex>
-    )
-  }
-
   return (
     <Modal {...modalProps}>
       <Flex flexDirection={DIRECTION_COLUMN}>
@@ -157,6 +108,7 @@ export function RunFailedModal({
               : []
           }
           isSingleError={!!highestPriorityError}
+          runStatus={runStatus}
         />
         <LegacyStyledText as="p">
           {t('branded:run_failed_modal_description_desktop')}
@@ -181,20 +133,3 @@ export function RunFailedModal({
     </Modal>
   )
 }
-
-const ERROR_MESSAGE_STYLE = css`
-  display: ${DISPLAY_FLEX};
-  flex-direction: ${DIRECTION_COLUMN};
-  max-height: 9.5rem;
-  overflow-y: ${OVERFLOW_AUTO};
-  margin-top: ${SPACING.spacing8};
-  margin-bottom: ${SPACING.spacing16};
-  padding: ${`${SPACING.spacing8} ${SPACING.spacing12}`};
-  background-color: ${COLORS.grey30};
-  border-radius: ${BORDERS.borderRadius8};
-  overflow-wrap: ${OVERFLOW_WRAP_ANYWHERE};
-
-  ::-webkit-scrollbar-thumb {
-    background: ${COLORS.grey40};
-  }
-`
